requirements/base.txt
click==8.0.3
elasticsearch[async]==7.8.0
fastapi==0.78.0
orjson==3.5.2
pydantic==1.8.2
python-jose==3.3.0
uvicorn==0.15.0
uvloop==0.16.0


